Change to call a diceroll for choosing which partition to switch to
